typos
Halo!

There's a typo at lines 59 and 114. I'm not sure if I'm right with the latter tho. I don't know how attention mechanisms work -- I just followed the output sizes of each operation. The edit shouldn't break the NMT codes since `hidden` is the same for both the encoder and the decoder.